// frontend/src/app/interfaces/enums/PackageStep.ts
export enum PackageStep {
  HOME = 'home',
  FLIGHT_DEPARTURE = 'flight_departure',
  FLIGHT_RETURN = 'flight_return',
  ACCOMMODATION = 'accommodation',
  ACCOMMODATION_DETAILS = 'accommodation_details',
  ACTIVITY = 'activity',
  ACTIVITY_DETAILS = 'activity_details',
  CHECKOUT = 'checkout',
  CONFIRM = 'confirm',
}